style: fix and pretty-format .vscode/launch.json file
This launch.json file is an autogenerated vscode-go file.
VSCode uses Json with Comments(jsonc) mode for its configuration files.

If we can find a way to introduce jsonc check to our pre-commit hooks,
we can change back the format of this file.

To read more about JsonC:
https://code.visualstudio.com/docs/languages/json#_json-with-comments

There is also an extension to JSON called Json5, Json for Humans.
https://json5.org/

Json5 formatter/checkers can be replaced with Json later.