fix: Fix a bug for credential data parsing in tfprotov5 resources
We're parsing the data to the `data` variable in this function, but then
the return api.NewAPI function uses only the schema data.

In this case, the credentials are not fetched from the env vars for the
tfprotov5 a.k.a. SDKv2 resources, such as biganimal_cluster.

because ba_bearer_token and ba_api_uri are not set for the Client, the
calls fail, such as

Error: Post "/projects/prj_deadbeef012345667/clusters": unsupported protocol scheme ""

This change fixes the issue.